refactor: Remove dependency on Boost.program_options from detectors (#1697)
This removes the dependencies of the Examples detectors from Boost.program_options. This is done by moving the class `IDetectorBase` to the `Run` folder and creating new classes called `GenericDetectorWithOptions` etc., that inherit from it.

The next step would be to make the binaries compilation optional, but I wanted to seperate these things from each other.