refactor!: Fitters use loggers more consistently (#1727)
First follow up to #1724 (includes that PR as well), which switches the
KF, CKF, GSF and GX2F to use `Logger` directly, getting rid of the
`LoggerWrapper`. This uses the new clone mechanism to create child
loggers for the actor.

To remove the logger instance from the propagation options, I had to
change all the actors and aborters, to take the logger as an explicit
argument. I changed the ActionList and AbortList to just forward any
arguments that are passed into the call operator, as that should be more
flexible. Since this makes the sort-of-concept-style call signature
checks very complex, I tentatively removed those as well now. I think
it's fine to drop these, they don't provide too much value here, and the
action and abort list are in need of an actual reimplementation anyway.